backend/requirements.txt
fastapi
uvicorn[standard]
python-multipart
openai
python-dotenv
sentence-transformers
pypdf
numpy
streamlit
requests
ollama
chromadb
